Move getTaxBN and bnMultiplyByFraction to util.js
